+++ exmg-paper-combobox.ts
@@ -1,24 +1,27 @@
 import {LitElement, html, customElement, property, query} from 'lit-element';
 import {classMap} from 'lit-html/directives/class-map';
 
-import '@polymer/paper-menu-button/paper-menu-button';
-import '@polymer/paper-listbox/paper-listbox';
-import '@polymer/paper-icon-button/paper-icon-button';
+import '@polymer/paper-menu-button/paper-menu-button.js';
+import '@polymer/paper-listbox/paper-listbox.js';
+import '@polymer/paper-icon-button/paper-icon-button.js';
 
-import '@polymer/iron-input/iron-input';
-import '@polymer/paper-input/paper-input-error';
-import '@polymer/paper-button/paper-button';
-import '@polymer/iron-flex-layout/iron-flex-layout';
-import '@polymer/paper-input/paper-input-container';
-import '@polymer/paper-styles/paper-styles';
-import './exmg-paper-combobox-icons';
-import {afterNextRender} from '@polymer/polymer/lib/utils/render-status';
+import '@polymer/iron-input/iron-input.js';
+import '@polymer/paper-input/paper-input-error.js';
+import '@polymer/paper-button/paper-button.js';
+import '@polymer/iron-flex-layout/iron-flex-layout.js';
+import '@polymer/paper-input/paper-input-container.js';
+import '@polymer/paper-styles/paper-styles.js';
+import './exmg-paper-combobox-icons.js';
+import {afterNextRender} from '@polymer/polymer/lib/utils/render-status.js';
 
-import {GenericPropertyValues, isEventWithPath, Token} from './exmg-custom-types';
-import {PaperMenuButton} from "@polymer/paper-menu-button/paper-menu-button"; /* It is repeated on purpose otherwise element won't works */ // tslint:disable-line
-import  {PaperListboxElement} from '@polymer/paper-listbox/paper-listbox'; /* It is repeated on purpose otherwise element won't works */ // tslint:disable-line
+import {EventSelectPayload, GenericPropertyValues, isEventWithPath, Token} from './exmg-custom-types';
+import {PaperMenuButton} from '@polymer/paper-menu-button/paper-menu-button';
+import  {PaperListboxElement} from '@polymer/paper-listbox/paper-listbox';
 
 type PrivateProps = 'inputValue' | 'selectedValue';
+type Props = Exclude<keyof PaperComboboxElement, number | Symbol> | PrivateProps;
+
+type ChangedProps = GenericPropertyValues<Props>;
 
 const copyElementStyle = (source: HTMLElement, target: HTMLElement): void  => {
   const computedStyle = window.getComputedStyle(source, null);
@@ -36,8 +39,20 @@ const copyElementStyle = (source: HTMLElement, target: HTMLElement): void  => {
 (window as any).Exmg = (window as any).Exmg || {};
 
 const BACKSPACE = 8;
-const ARROWDOWN = 40;
+const ARROW_DOWN = 40;
 const DELETE = 127;
+const ESC = 27;
+
+const debounce  = (time: number) => {
+  let timer: number;
+
+  return (cb?: Function): void => {
+    clearTimeout(timer);
+    if (cb) {
+      timer = window.setTimeout(cb, time);
+    }
+  };
+};
 
 /**
 * `exmg-paper-combobox` is an wrapper element to make list data selectable.
@@ -66,8 +81,8 @@ const DELETE = 127;
 * @memberof Exmg
 * @extends LitElement
 * @summary Paper Combobox Element
-* @eventType exmg-combobox-select - detail is equal to this.selected value
-* @eventType exmg-combobox-deselect - detail is equal to undefined
+* @eventType exmg-combobox-select - detail is type EventSelectPayload
+* @eventType exmg-combobox-deselect - detail undefined
 */
 @customElement('exmg-paper-combobox')
 export class PaperComboboxElement extends LitElement {
@@ -95,15 +110,21 @@ export class PaperComboboxElement extends LitElement {
    * Returns currently selected item.
    * @type {?Object}
    */
-  @property({type: Object, attribute: 'selected-item'}) selectedItem?: Element; // @todo notify
+  @property({type: Object, attribute: 'selected-item'}) selectedItem?: Element;
 
   /**
    * Gets or sets the selected element. The default is to use the index of the item.
    * @type {string|number}
    */
-  @property({type: String}) selected?: string | number; // @todo notify: true, observe: _observeSelected
+  @property({type: String}) selected?: string | number;
 
-  @property({type: String, attribute: 'selected-value'}) private selectedValue?: string | number; // @todo observe: _observeSelectedValue
+  /**
+   * Set custom max width of menu list with items
+   * @type {number = 200}
+   */
+  @property({type: Number, attribute: 'max-width-menu-list'}) maxWidthMenuList: number = 200;
+
+  @property({type: String}) private selectedValue?: string | number;
 
   /**
    * The label for this input.
@@ -174,12 +195,21 @@ export class PaperComboboxElement extends LitElement {
 
   private ignoreFocus: boolean = false;
 
-  private observers = this.getObservers();
+  private isAnyItemActive: boolean = true;
+
+  private isElementInitialized: boolean = false;
+
+  private readonly observers: {[K in Props]?: Function} = this.getObservers();
+
+  private readonly keyDownBackspaceDebounce: (cb?: Function) => void = debounce(200);
+
+  private readonly inputChangeDebounce: (cb?: Function) => void  = debounce(300);
 
   constructor() {
     super();
     this.onKeyDown = this.onKeyDown.bind(this);
     this.onClick = this.onClick.bind(this);
+    this.onIronResize = this.onIronResize.bind(this);
   }
 
   /***************** OBSERVERS *******************/
@@ -187,7 +217,7 @@ export class PaperComboboxElement extends LitElement {
   /**
    * Register observed properties and actions to perform
    */
-  private getObservers(): Record<keyof this, Function> | Record<PrivateProps, Function> {
+  private getObservers(): {[K in Props]?: Function} {
     return {
       inputValue: () => this.observeInputChange(),
       selectedItem: () => this.observeSelectedItem(),
@@ -196,10 +226,10 @@ export class PaperComboboxElement extends LitElement {
     };
   }
 
-  private executeObservers(changedProperties: GenericPropertyValues<keyof this | PrivateProps>): void {
+  private executeObservers(changedProperties: ChangedProps): void {
     Object.entries(this.observers).forEach(([key, cb]) => {
-      if (changedProperties.has(<any>key)) {
-        cb();
+      if (cb && changedProperties.has(<Props>key)) {
+        cb(changedProperties);
       }
     });
   }
@@ -209,7 +239,24 @@ export class PaperComboboxElement extends LitElement {
       this.inputElement.style.width = `${(this.inputWidthHelperElement.offsetWidth + 10)}px`;
     }
 
-    this.filterItems();
+    if (!this.isElementInitialized) {
+      this.isAnyItemActive = this.filterItems();
+    }
+
+    if (this.isElementInitialized && this.menuElement) {
+      this.inputChangeDebounce(() => {
+        this.isAnyItemActive = this.filterItems();
+        this.onIronResize();
+
+        if (!this.menuElement!.opened && this.isAnyItemActive && !!this.inputValue) {
+          this.menuElement!.open();
+          afterNextRender(this, () => this.focus());
+        } else if (this.menuElement!.opened && !this.isAnyItemActive) {
+          this.menuElement!.close();
+        }
+
+      });
+    }
   }
 
   private observeSelectedValue(): void {
@@ -247,17 +294,22 @@ export class PaperComboboxElement extends LitElement {
     this.token = {id, text};
   }
 
-  filterItems() {
+  filterItems(): boolean {
     const items: NodeListOf<HTMLElement> = this.querySelectorAll('paper-item, paper-icon-item');
     const hasFilterPhrase = !!this.inputValue && this.inputValue.length > 0;
+    const phrase = hasFilterPhrase ? this.inputValue.toLowerCase().trim() : '';
+    let isAnyItemActive = false;
 
     items.forEach(item => {
-      if (hasFilterPhrase && item.textContent &&  item.textContent.indexOf(this.inputValue || '') === -1) {
+      if (hasFilterPhrase && item.textContent &&  item.textContent.toLowerCase().indexOf(phrase) === -1) {
         item.setAttribute('hidden', '');
       } else {
+        isAnyItemActive = true;
         item.removeAttribute('hidden');
       }
     });
+
+    return isAnyItemActive;
   }
 
   indexOf(item: any): number {
@@ -265,9 +317,13 @@ export class PaperComboboxElement extends LitElement {
   }
 
   private getSelectedItemKey(selectedItem: Element): number | string | undefined {
-    return this.attrForSelected ?
-        selectedItem.getAttribute(this.attrForSelected) || undefined :
-        this.indexOf(selectedItem);
+    if (!!this.attrForSelected) {
+      return  selectedItem.getAttribute(this.attrForSelected) || undefined;
+    }
+
+    const index = this.indexOf(selectedItem);
+
+    return index === -1 ? undefined : index;
   }
 
   private hasSelectedItem(): boolean {
@@ -289,7 +345,7 @@ export class PaperComboboxElement extends LitElement {
       return true;
     }
 
-    return !(!this.token && this.inputElement !== document.activeElement);
+    return !!this.token || this.inputFocused;
   }
 
   private resetInput(): void {
@@ -298,6 +354,7 @@ export class PaperComboboxElement extends LitElement {
     }
 
     this.inputValue = '';
+    this.inputElement!.value = '';
     if (this.ignoreFocus) {
       this.ignoreFocus = false;
     } else {
@@ -310,33 +367,50 @@ export class PaperComboboxElement extends LitElement {
    * @return {boolean} True if the value is valid.
    */
   validate(): boolean {
-    this.invalid = this.required && !this.hasSelectedItem();
+    this.invalid = !this.disabled && this.required && !this.hasSelectedItem();
     return !this.invalid;
   }
 
   /************ EVENT HANDLERS *************/
   private onKeyDown(e: KeyboardEvent): void {
-    if (!this.inputValue) {
+    if (typeof this.inputValue !== 'string') {
       this.inputValue = '';
     }
 
-    switch (e.keyCode) {
+    switch (e.code || e.keyCode) {
       case BACKSPACE:
+      case 'Backspace':
       case DELETE:
-        this.opened = true;
+      case 'Delete':
+        if (!this.menuElement!.opened) {
+          this.keyDownBackspaceDebounce(() => {
+            if (!this.menuElement!.opened && this.isAnyItemActive) {
+              this.menuElement!.open();
+              afterNextRender(this, () => this.focus());
+            }
+          });
+        }
+
         this.selectedValue = undefined;
         this.selectedItem = undefined;
+        break;
+      case ARROW_DOWN:
+      case 'ArrowDown':
+        if (!this.menuElement!.opened && this.isAnyItemActive) {
+          this.menuElement!.open();
+        }
+        afterNextRender(this, () => this.listBox!.focus());
+        break;
+      case ESC:
+      case 'Escape':
+        e.preventDefault();
+        this.menuElement!.close();
         afterNextRender(this, () => this.focus());
-        break;
-      case ARROWDOWN:
-        this.opened = true;
-        this.listBox!.focus();
-        break;
     }
   }
 
   private onClick(e: Event): void {
-    const inside: boolean = isEventWithPath(e) ? !!e.path && !!e.path.find((path) => path === this) : false;
+    const inside: boolean = isEventWithPath(e) ? !!e.path && !!e.composedPath().find((path) => path === this) : e.target === this;
 
     // Detect outside element click for auto validate input
     if (this.autoValidate && (this.previousInsideClick && !inside) || this.token) {
@@ -358,6 +432,23 @@ export class PaperComboboxElement extends LitElement {
     this.selected = this.getSelectedItemKey(this.selectedItem);
 
     this.resetInput();
+  }
+
+  private onItemDeselected(e: Event): void {
+    e.stopPropagation();
+    this.selectedItem = undefined;
+    this.selected = undefined;
+
+    this.resetInput();
+  }
+
+  private onItemActivated(e: CustomEvent<{selected: string | number}>): void {
+    // when user select same item then don't receive iron-select event but we still want to
+    // prepare input
+    if (this.selected === e.detail.selected) {
+      this.ignoreFocus = false;
+      afterNextRender(this, () => this.resetInput());
+    }
   }
 
   private onMenuButtonOpen(e: Event): void {
@@ -401,29 +492,80 @@ export class PaperComboboxElement extends LitElement {
     if (this.autoValidate) {
       window.addEventListener('click', this.onClick);
     }
+
+    this.addEventListener('iron-resize', this.onIronResize);
+  }
+
+  /**
+   * Fix menu content width and height
+   */
+  private onIronResize(): void {
+    const element: HTMLElement = this.menuElement!.shadowRoot!.querySelector<HTMLElement>('.dropdown-content')!;
+
+    const {left: elementLeft} = element.getBoundingClientRect();
+    const {scrollWidth: elementScrollWidth} = element;
+    const getGreater = (...values: number[]): number  => Math.max(...values);
+    const getLower = (...values: number[]): number  => Math.min(...values);
+
+    if (elementScrollWidth > 0 && elementScrollWidth < this.maxWidthMenuList) {
+      const elementMaximumWidthFromRight = document.documentElement.clientWidth - elementLeft;
+      element.style.maxWidth = `${getLower(getGreater(elementMaximumWidthFromRight, 100), this.maxWidthMenuList)}px`;
+    }
+
+    const {top:  elementTop} = element.getBoundingClientRect();
+    const {scrollHeight: elementScrollHeight} = element;
+    if (elementScrollHeight > 0) {
+      const elementMaximumHeightToBottom = document.documentElement.clientHeight - elementTop;
+      element.style.maxHeight = `${getGreater(elementMaximumHeightToBottom - 10, 100)}px`;
+    }
+  }
+
+  private shouldFireEvent(changedProperties: GenericPropertyValues<keyof this | PrivateProps>): boolean {
+    const props: (keyof this | PrivateProps)[] = ['selected', 'selectedItem'];
+    const anyPropChanged = props.some((it: keyof this | PrivateProps) =>
+      // @ts-ignore
+      changedProperties.has(it) && changedProperties.get(it) !== this[it]
+    );
+
+    const {id = undefined} = this.token || {};
+
+    return (anyPropChanged && id === this.selected);
   }
 
   /*****  LIT ELEMENT HOOKS ******/
 
-  protected firstUpdated(): void {
+  protected async firstUpdated(): Promise<void> {
     this.initializeElement();
+    await this.updateComplete;
+    this.isElementInitialized = true;
   }
 
-  protected updated(changedProperties: GenericPropertyValues<keyof this | PrivateProps>) {
+  protected updated(changedProperties: ChangedProps) {
     this.executeObservers(changedProperties);
+    if (this.shouldFireEvent(changedProperties)) {
+      if (typeof this.selected !== 'undefined') {
+        const payload: EventSelectPayload = {
+          value: this.selected!,
+          item: this.selectedItem!,
+          token: this.token!,
+        };
 
-    if (changedProperties.has('selected') && changedProperties.get('selected') !== this.selected) {
-      const eventName = (typeof this.selected === 'undefined' || this.selected === -1) ? 'exmg-combobox-deselect' : 'exmg-combobox-select';
-      this.dispatchEvent(new CustomEvent(eventName, {detail: this.selected, composed: true, bubbles: true}));
+        this.dispatchEvent(new CustomEvent('exmg-combobox-select', {detail: payload, composed: true, bubbles: true}));
+      } else {
+        this.dispatchEvent(new CustomEvent('exmg-combobox-deselect', {composed: true, bubbles: true}));
+      }
     }
   }
 
   disconnectedCallback() {
     super.disconnectedCallback();
     this.inputElement && this.inputElement.removeEventListener('keydown', this.onKeyDown);
+    this.removeEventListener('iron-resize', this.onIronResize);
     if (this.autoValidate) {
       window.removeEventListener('click', this.onClick);
     }
+    this.inputChangeDebounce();
+    this.keyDownBackspaceDebounce();
   }
 
   protected render() {
@@ -550,8 +692,9 @@ export class PaperComboboxElement extends LitElement {
             attr-for-selected="${this.attrForSelected}"
             selected="${this.selectedValue}"
             class="dropdown-content"
+            @iron-activate="${this.onItemActivated}"
             @iron-select="${this.onItemSelected}"
-            @iron-deselect="${this.resetInput}">
+            @iron-deselect="${this.onItemDeselected}">
             <slot></slot>
           </paper-listbox>
         </paper-dropdown-menu>