+++ CHANGELOG.md
@@ -20,7 +20,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Rewrite component to LitElement
 ### Changed
 * `[Breaking]` Two way data binding has been removed. To obtain `selected` value 2 handlers should be added:
-  * event `exmg-combobox-select` where event.detail is equal to `element.selected`
+  * event `exmg-combobox-select` where event.detail is equal to `EventSelectPayload`
+  ```typescript
+    export type EventSelectPayload = {
+      value: number | string;
+      item: Element;
+      token: Token;
+    };
+    
+    export type Token = {
+        id: number | string;
+        text: string;
+    };
+  ```
   * event `exmg-combobox-delect`
   * example for lit-element
   ```html